advocate good file structure
