+++ .prettierignore
@@ -11,3 +11,5 @@ packages/tests/server/___packages.ts
 packages/tests/vendor/
 packages/coverage/
 pnpm-lock.yaml
+
+www/docs/main/quickstart.mdx
