+++ examples/minimal/server/index.ts
@@ -1,27 +1,7 @@
-import { initTRPC } from '@trpc/server';
 import { createHTTPServer } from '@trpc/server/adapters/standalone';
 import { z } from 'zod';
-
-const t = initTRPC.create();
-
-const router = t.router;
-const publicProcedure = t.procedure;
-
-type User = { id: string; name: string };
-
-// Imaginary database
-const users: User[] = [];
-const db = {
-  user: {
-    findMany: async () => users,
-    findById: async (id: string) => users.find((user) => user.id === id),
-    create: async (data: { name: string }) => {
-      const user = { id: String(users.length + 1), ...data };
-      users.push(user);
-      return user;
-    },
-  },
-};
+import { db } from './db';
+import { publicProcedure, router } from './trpc';
 
 const appRouter = router({
   userList: publicProcedure.query(async () => {
