+++ examples/minimal/server/trpc.ts
@@ -1,0 +1,14 @@
+import { initTRPC } from '@trpc/server';
+
+/**
+ * Initialization of tRPC backend
+ * Should be done only once per backend!
+ */
+const t = initTRPC.create();
+
+/**
+ * Export reusable router and procedure helpers
+ * that cane be used throughout the router
+ */
+export const router = t.router;
+export const publicProcedure = t.procedure;
